first refactor commit
