+++ data/sites/no_websites.csv
@@ -1,0 +1,2 @@
+url
+this is not a URI
